requirements.txt
agno
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
fastapi
python-multipart
pgvector
psycopg[binary]
pypdf